Rename and move health check to python
