+++ python/health-check/src/health_check/config/grafana/dashboards/dashboard_with_logs.json
