+++ python/health-check/src/health_check/config/templates/grafana_dashboard/supportconfig_with_logs.template.json
