+++ python/health-check/src/health_check/config/templates/promtail/promtail.yaml.j2
