+++ python/health-check/src/health_check/containers/exporter/Dockerfile
