+++ python/health-check/src/health_check/containers/loki/Dockerfile
