+++ python/health-check/src/health_check/exporters/static_metrics.py
